added dynamic memory allocation for number of alignments

+++ src/main.cpp
@@ -69,7 +69,7 @@ bool full_search_gv = false;
 char version_num[] = "2.0-dev, 29/11/2014";
 bool as_percent_gv = false;
 bool pid_gv = false;
-int16_t num_best_hits_gv = 0;
+int32_t num_best_hits_gv = 0;
